==> notebooks/03_stemming_lemmatization.ipynb <==
# Modules import

import os.path

import pandas as pd
import nltk
from nltk.tokenize import TweetTokenizer
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer, WordNetLemmatizer

nltk.download('stopwords')
nltk.download('wordnet')

# Data import

DATA_PATH = os.path.join('..', 'data', 'preprocessed_url_simple')
STOPWORDS = set(stopwords.words('english'))
train = pd.read_csv(os.path.join(DATA_PATH, 'train.csv'))
train.head()

validation = pd.read_csv(os.path.join(DATA_PATH, 'validation.csv'))
validation.head()

test = pd.read_csv(os.path.join(DATA_PATH, 'test.csv'))
test.head()

# Data preprocessing
## Tokenization And Stopwords removal

def tokenize_remove_stop_words(dataset: pd.DataFrame) -> pd.DataFrame:
    """
    Tokenize text and remove stopwords
    :param dataset: Dataset to process
    :return: Processed dataset
    """
    tknzr = TweetTokenizer(strip_handles=True)
    dataset['tokens'] = dataset['text'].apply(lambda x: tknzr.tokenize(x))
    dataset['tokens'] = dataset['tokens'].apply(lambda x: [w for w in x if not w in STOPWORDS])
    return dataset


train = tokenize_remove_stop_words(train)
train.head()

validation = tokenize_remove_stop_words(validation)
validation.head()

test = tokenize_remove_stop_words(test)
test.head()

## Stemming

PS = PorterStemmer()


def stemming(dataset: pd.DataFrame) -> pd.DataFrame:
    """
    Stemming text
    :param dataset: Dataset to process 
    :return: Processed dataset
    """
    new_dataset = dataset.copy()
    new_dataset['tokens'] = new_dataset['tokens'].apply(lambda x: [PS.stem(w) for w in x])
    new_dataset['new_text'] = new_dataset['tokens'].apply(lambda x: ' '.join(x))
    return new_dataset


stemmed_train = stemming(train)
stemmed_validation = stemming(validation)
stemmed_test = stemming(test)
stemmed_train.head()

### Saving stemmed data

STEMMED_DATA_PATH = os.path.join('..', 'data', 'stemmed')
if not os.path.exists(STEMMED_DATA_PATH):
    os.mkdir(STEMMED_DATA_PATH)
stemmed_train.to_csv(os.path.join(STEMMED_DATA_PATH, 'train.csv'), index=False)
stemmed_validation.to_csv(os.path.join(STEMMED_DATA_PATH, 'validation.csv'), index=False)
stemmed_test.to_csv(os.path.join(STEMMED_DATA_PATH, 'test.csv'), index=False)

## Lemmatization

LEMMATIZER = WordNetLemmatizer()


def lemmatization(dataset: pd.DataFrame) -> pd.DataFrame:
    """
    Lemmatization of text
    :param dataset: Dataset to process
    :return: Processed dataset
    """
    new_dataset = dataset.copy()
    new_dataset['tokens'] = new_dataset['tokens'].apply(lambda x: [LEMMATIZER.lemmatize(w) for w in x])
    new_dataset['new_text'] = new_dataset['tokens'].apply(lambda x: ' '.join(x))
    return new_dataset


lemmatized_train = lemmatization(train)
lemmatized_validation = lemmatization(validation)
lemmatized_test = lemmatization(test)

### Saving lemmatized data

LEMMATIZED_DATA_PATH = os.path.join('..', 'data', 'lemmatized')
if not os.path.exists(LEMMATIZED_DATA_PATH):
    os.mkdir(LEMMATIZED_DATA_PATH)
stemmed_train.to_csv(os.path.join(LEMMATIZED_DATA_PATH, 'train.csv'), index=False)
stemmed_validation.to_csv(os.path.join(LEMMATIZED_DATA_PATH, 'validation.csv'), index=False)
stemmed_test.to_csv(os.path.join(LEMMATIZED_DATA_PATH, 'test.csv'), index=False)
