==> notebooks/11_modeling_tfidf.ipynb <==
# Modules import

import os.path
import pickle

import numpy as np
import optuna
import pandas as pd
from sklearnex import patch_sklearn

patch_sklearn()
import utils.optuna_utils as ou
from optuna import create_study

optuna.logging.set_verbosity(optuna.logging.WARNING)

TIMEOUT = 1200  # seconds
TRIALS = 100  # number of trials
N_SPLITS = 5

np.random.seed(42)

# Simple processing

DATA_PATH = os.path.join('..', 'data', 'preprocessed_url_simple')
train = pd.read_csv(os.path.join(DATA_PATH, 'train.csv'))
validation = pd.read_csv(os.path.join(DATA_PATH, 'validation.csv'))
test = pd.read_csv(os.path.join(DATA_PATH, 'test.csv'))
train.head()

def get_x_y(df: pd.DataFrame) -> (pd.Series, pd.Series):
    """
    Get predictors and target from dataframe
    :param df: dataframe
    :return: predictors, target
    """
    x = df["text"]
    y = df["account.type"]
    y = np.where(y == "bot", 1, 0)
    return x, y


x_train, y_train = get_x_y(train)
x_validation, y_validation = get_x_y(validation)
x_train = pd.concat([x_train, x_validation])
y_train = np.append(y_train, y_validation)
x_train.shape, y_train.shape

x_test, y_test = get_x_y(test)
x_test.shape, y_test.shape

preprocessed_path = os.path.join("..", "models", "tfidf", "preprocessed")
os.makedirs(preprocessed_path, exist_ok=True)

studies = []
results_optuna = []
results_test = []
for model in ou.MODELS.keys():
    study = create_study(study_name=f'simple_processing_{model}', direction='maximize')
    study.optimize(lambda trial: ou.objective(trial, x_train, y_train, model=model, encoder="TFIDF", n_splits=N_SPLITS),
                   timeout=TIMEOUT, show_progress_bar=True, n_trials=TRIALS)
    studies.append(study)
    retrained_model = ou.get_best_model(study.best_params, x_train, y_train)
    with open(os.path.join(preprocessed_path, f"{model}.pickle"), "wb") as f:
        pickle.dump(retrained_model, f)
    results_test.append({"type": "preprocessed",
                         "model": model,
                         **ou.get_score(retrained_model, x_test, y_test)})
    print(f"Model: {model}")
    print(f"Best accuracy: {study.best_value}")
    print(f"Best params: {study.best_params}")
    results_optuna.append({
        "type": "preprocessed",
        "model": model,
        "score": study.best_value,
        "params": study.best_params
    })

# Stemming

DATA_PATH = os.path.join('..', 'data', 'stemmed')
train = pd.read_csv(os.path.join(DATA_PATH, 'train.csv'))
validation = pd.read_csv(os.path.join(DATA_PATH, 'validation.csv'))
test = pd.read_csv(os.path.join(DATA_PATH, 'test.csv'))
train.head()

x_train, y_train = get_x_y(train)
x_validation, y_validation = get_x_y(validation)
x_train = pd.concat([x_train, x_validation])
y_train = np.append(y_train, y_validation)
x_train.shape, y_train.shape

x_test, y_test = get_x_y(test)
x_test.shape, y_test.shape

preprocessed_path = os.path.join("..", "models", "tfidf", "stemmed")
os.makedirs(preprocessed_path, exist_ok=True)

for model in ou.MODELS.keys():
    study = create_study(study_name=f'stemming_{model}', direction='maximize')
    study.optimize(lambda trial: ou.objective(trial, x_train, y_train, model=model, encoder="TFIDF", n_splits=N_SPLITS),
                   timeout=TIMEOUT, show_progress_bar=True, n_trials=TRIALS)
    studies.append(study)
    retrained_model = ou.get_best_model(study.best_params, x_train, y_train)
    with open(os.path.join(preprocessed_path, f"{model}.pickle"), "wb") as f:
        pickle.dump(retrained_model, f)
    results_test.append({"type": "stemming",
                         "model": model,
                         **ou.get_score(retrained_model, x_test, y_test)})
    print(f"Model: {model}")
    print(f"Best accuracy: {study.best_value}")
    print(f"Best params: {study.best_params}")
    results_optuna.append({
        "type": "stemming",
        "model": model,
        "score": study.best_value,
        "params": study.best_params
    })

# Lemmatization

DATA_PATH = os.path.join('..', 'data', 'lemmatized')
train = pd.read_csv(os.path.join(DATA_PATH, 'train.csv'))
validation = pd.read_csv(os.path.join(DATA_PATH, 'validation.csv'))
test = pd.read_csv(os.path.join(DATA_PATH, 'test.csv'))
train.head()

x_train, y_train = get_x_y(train)
x_validation, y_validation = get_x_y(validation)
x_train = pd.concat([x_train, x_validation])
y_train = np.append(y_train, y_validation)
x_train.shape, y_train.shape

x_test, y_test = get_x_y(test)
x_test.shape, y_test.shape

preprocessed_path = os.path.join("..", "models", "tfidf", "lemmatization")
os.makedirs(preprocessed_path, exist_ok=True)

for model in ou.MODELS.keys():
    study = create_study(study_name=f'lemmatization_{model}', direction='maximize')
    study.optimize(lambda trial: ou.objective(trial, x_train, y_train, model=model, encoder="TFIDF", n_splits=N_SPLITS),
                   timeout=TIMEOUT, show_progress_bar=True, n_trials=TRIALS)
    studies.append(study)
    retrained_model = ou.get_best_model(study.best_params, x_train, y_train)
    with open(os.path.join(preprocessed_path, f"{model}.pickle"), "wb") as f:
        pickle.dump(retrained_model, f)
    results_test.append({"type": "lemmatization",
                         "model": model,
                         **ou.get_score(retrained_model, x_test, y_test)})
    print(f"Model: {model}")
    print(f"Best accuracy: {study.best_value}")
    print(f"Best params: {study.best_params}")
    results_optuna.append({
        "type": "lemmatization",
        "model": model,
        "score": study.best_value,
        "params": study.best_params
    })

# Results

results_df = pd.DataFrame(results_optuna)
results_df.sort_values(by="score", ascending=False)

results_df.sort_values(by="score", ascending=False).to_csv(
    os.path.join("results", "tfidf_optuna.csv"), index=False)

results_test_df = pd.DataFrame(results_test)
results_test_df.sort_values(by="balanced_accuracy", ascending=False)

results_test_df.sort_values(by="balanced_accuracy", ascending=False).to_csv(
    os.path.join("results", "tfidf.csv"), index=False)
