==> notebooks/33_WordCNN+GRU.ipynb <==
# Imports

import os

import keras
import pandas as pd

import utils.word_utils as wu
from utils.KerasModels import calculate_metrics

keras.utils.set_random_seed(812)
MODELS_DIR = os.path.join('..', 'models', 'wordCNN_GRU')
EPOCHS = 200
MAX_LEN = 100
VOCAB_SIZE = 15000

# Simple data

data_path = os.path.join('..', 'data', 'preprocessed_url_simple')

x_train, y_train, x_valid, y_valid, x_test, y_test = wu.prepare_data(
    data_path)
os.makedirs(MODELS_DIR, exist_ok=True)
text_vectorizer = wu.prepare_text_vectorizer(x_train, max_len=MAX_LEN, vocab_size=VOCAB_SIZE)
model = wu.get_wordcnn_gru(text_vectorizer)
model.summary()

# Train CNN model.
early_stopping = keras.callbacks.EarlyStopping(monitor='val_loss', patience=10, restore_best_weights=True)
save_best = keras.callbacks.ModelCheckpoint(os.path.join(MODELS_DIR, "simple"), monitor='val_loss',
                                            save_best_only=True, restore_best_weights=True)
model.fit(x_train, y_train, batch_size=256, epochs=EPOCHS,
          validation_data=(x_valid, y_valid),
          callbacks=[early_stopping, save_best])

probas = model.predict(x_test)
y_pred = (probas > 0.5).astype('int32')
results = calculate_metrics(y_test, y_pred)
results

# Lemmatized data

data_path = os.path.join('..', 'data', 'lemmatized')

x_train, y_train, x_valid, y_valid, x_test, y_test = wu.prepare_data(
    data_path)
os.makedirs(MODELS_DIR, exist_ok=True)
text_vectorizer = wu.prepare_text_vectorizer(x_train, max_len=MAX_LEN, vocab_size=VOCAB_SIZE)
model = wu.get_wordcnn_gru(text_vectorizer)
model.summary()

# Train CNN model.
early_stopping = keras.callbacks.EarlyStopping(monitor='val_loss', patience=10, restore_best_weights=True)
save_best = keras.callbacks.ModelCheckpoint(os.path.join(MODELS_DIR, "lemmatized"), monitor='val_loss',
                                            save_best_only=True, restore_best_weights=True)
model.fit(x_train, y_train, batch_size=256, epochs=EPOCHS,
          validation_data=(x_valid, y_valid),
          callbacks=[early_stopping, save_best])

probas = model.predict(x_test)
y_pred = (probas > 0.5).astype('int32')
results_lem = calculate_metrics(y_test, y_pred)
results_lem

# Stemmed data

data_path = os.path.join('..', 'data', 'stemmed')

x_train, y_train, x_valid, y_valid, x_test, y_test = wu.prepare_data(
    data_path)
os.makedirs(MODELS_DIR, exist_ok=True)
text_vectorizer = wu.prepare_text_vectorizer(x_train, max_len=MAX_LEN, vocab_size=VOCAB_SIZE)
model = wu.get_wordcnn_gru(text_vectorizer)
model.summary()

# Train CNN model.
early_stopping = keras.callbacks.EarlyStopping(monitor='val_loss', patience=10, restore_best_weights=True)
save_best = keras.callbacks.ModelCheckpoint(os.path.join(MODELS_DIR, "stemmed"), monitor='val_loss',
                                            save_best_only=True, restore_best_weights=True)
model.fit(x_train, y_train, batch_size=256, epochs=EPOCHS,
          validation_data=(x_valid, y_valid),
          callbacks=[early_stopping, save_best])

probas = model.predict(x_test)
y_pred = (probas > 0.5).astype('int32')
results_stem = calculate_metrics(y_test, y_pred)
results_stem

# Results

# Create csv with all results

results_all = pd.DataFrame([results, results_lem, results_stem])
results_all['model'] = 'WordCNN+GRU'
results_all['dataset'] = ['simple', 'lemmatized', 'stemmed']

if not os.path.exists('results'):
    os.makedirs('results')
results_all.to_csv('results/word_cnn_gru.csv', index=False)
results_all
