==> notebooks/42_2-DistilBert+gpt-output.ipynb <==
import os

import random
import numpy as np
import pandas as pd

from transformers import AutoTokenizer

# from transformers import XLMRobertaForSequenceClassification, AdamW, get_linear_schedule_with_warmup
from transformers import (
    AutoModelForSequenceClassification,
    XLMRobertaForSequenceClassification,
    TrainingArguments,
    Trainer,
    EarlyStoppingCallback,
)

import json


import torch
import torch.nn as nn
from torch.optim import AdamW, Adam
from torch.utils.data import Dataset, DataLoader

seed = 1
model_checkpoint = "distilbert-base-uncased-full-training-gpt2-data"
model_name = "distilbert-base-uncased"
tokenizer_name = "distilbert-base-uncased"

model = AutoModelForSequenceClassification.from_pretrained(model_name, num_labels=2)
# model = AutoModelForSequenceClassification.from_pretrained(
#     f"{model_checkpoint}/checkpoint-2589", num_labels=2
# )

tokenizer = AutoTokenizer.from_pretrained(tokenizer_name)

model

def freeze_params(model):
    for name, param in model.named_parameters():
        if not name.startswith("classifier"):
            param.requires_grad = False

def seed_everything(seed):
    random.seed(seed)
    os.environ["PYTHONHASHSEED"] = str(seed)
    np.random.seed(seed)
    torch.manual_seed(seed)

seed_everything(1)

# freeze_params(model)
if torch.cuda.is_available():
    model = model.to("cuda")

train = pd.read_csv("data/original/train.csv")
test = pd.read_csv("data/original/test.csv")
valid = pd.read_csv("data/original/validation.csv")

LABEL_REAL = 0
LABEL_FAKE = 1


def load_texts(data_file: str) -> list[str]:
    texts = []

    with open(data_file) as f:
        for line in f:
            texts.append(json.loads(line)["text"])

    return texts


def build_pandas(
    fake: list[str],
    real: list[str],
    label_fake: str = LABEL_FAKE,
    label_real: str = LABEL_REAL,
) -> pd.DataFrame:
    df = pd.DataFrame(
        {
            "text": [*fake, *real],
            "label": np.concatenate(
                [[label_fake] * len(fake), [label_real] * len(real)]
            ),
        }
    )
    return df

train_fake = load_texts(
    os.path.join("data/gpt-2-ouput-dataset/medium-345M-k40.train.jsonl")
)
validation_fake = load_texts(
    os.path.join(f"data/gpt-2-ouput-dataset/medium-345M-k40.valid.jsonl")
)
test_fake = load_texts(
    os.path.join(f"data/gpt-2-ouput-dataset/medium-345M-k40.test.jsonl")
)
train_real = load_texts(os.path.join("data/gpt-2-ouput-dataset/webtext.train.jsonl"))
validation_real = load_texts(
    os.path.join("data/gpt-2-ouput-dataset/webtext.valid.jsonl")
)
test_real = load_texts(os.path.join("data/gpt-2-ouput-dataset/webtext.test.jsonl"))

train_gpt = build_pandas(train_fake, train_real)
valid_gpt = build_pandas(validation_fake, validation_real)
test_gpt = build_pandas(test_fake, test_real)

mapping = {"human": 0, "bot": 1}

train["label"] = train["account.type"].apply(lambda x: mapping[x])
test["label"] = test["account.type"].apply(lambda x: mapping[x])
valid["label"] = valid["account.type"].apply(lambda x: mapping[x])

class TweetDataset(Dataset):
    def __init__(self, data, tokenizer, max_len):
        self.data = data
        self.tokenizer = tokenizer
        self.max_len = max_len

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):
        data = self.data.iloc[index]

        text = data.text
        label = data.label

        encoding = self.tokenizer(
            text,
            add_special_tokens=True,
            max_length=self.max_len,
            return_token_type_ids=False,
            padding="max_length",
            truncation=True,
            return_attention_mask=True,
            return_tensors="pt",
        )

        return dict(
            text=text,
            input_ids=encoding["input_ids"].flatten(),
            attention_mask=encoding["attention_mask"].flatten(),
            label=torch.tensor(label, dtype=torch.long),
        )

batch_size = 8
lr = 2e-6

ds_train = TweetDataset(train, tokenizer=tokenizer, max_len=512)
ds_test = TweetDataset(test, tokenizer=tokenizer, max_len=512)
ds_valid = TweetDataset(valid, tokenizer=tokenizer, max_len=512)
ds_train_gpt = TweetDataset(train_gpt, tokenizer=tokenizer, max_len=512)
ds_test_gpt = TweetDataset(test_gpt, tokenizer=tokenizer, max_len=512)
ds_valid_gpt = TweetDataset(valid_gpt, tokenizer=tokenizer, max_len=512)

def compute_metrics(eval_pred):
    logits, labels = eval_pred
    predictions = np.argmax(logits, axis=-1)
    return {"accuracy": (predictions == labels).astype(np.float32).mean().item()}

args = TrainingArguments(
    f"{model_checkpoint}",
    evaluation_strategy="steps",
    save_strategy="steps",
    save_steps=2_000,
    eval_steps=2_000,
    num_train_epochs=1,
    per_device_train_batch_size=batch_size,
    per_device_eval_batch_size=batch_size,
    weight_decay=0.01,
    save_total_limit=2,  # save best and last
    load_best_model_at_end=True,
    seed=seed,
    # metric_for_best_model=metric_name,
)

trainer = Trainer(
    model,
    args,
    train_dataset=ds_train_gpt,
    eval_dataset=ds_valid_gpt,
    tokenizer=tokenizer,
    compute_metrics=compute_metrics,
    optimizers=(AdamW(model.parameters(), lr=lr), None),  # Optimizer, Scheduler
    # compute_loss to override one must create cutsom trainer class and override the method compute_loss(self, model, inputs, return_outputs=False):
    callbacks=[EarlyStoppingCallback(early_stopping_patience=5)],
)

trainer.train()

test_results = trainer.predict(ds_test_gpt)

test_results.metrics

from sklearn.metrics import (
    balanced_accuracy_score,
    f1_score,
    precision_score,
    recall_score,
)


def proba_to_pred(proba):
    pred = (proba > 0.5).astype(int)
    return pred


def calculate_metrics(y_true, y_pred):
    results = {
        "balanced_accuracy": balanced_accuracy_score(y_true, y_pred),
        "f1_score": f1_score(y_true, y_pred),
        "precision": precision_score(y_true, y_pred),
        "recall": recall_score(y_true, y_pred),
    }
    return results

calculate_metrics(test_gpt.label.values, np.argmax(test_results.predictions, axis=1))

args = TrainingArguments(
    f"{model_checkpoint}-with-twitter",
    evaluation_strategy="epoch",
    save_strategy="epoch",
    num_train_epochs=10,
    per_device_train_batch_size=batch_size,
    per_device_eval_batch_size=batch_size,
    weight_decay=0.01,
    save_total_limit=2,  # save best and last
    load_best_model_at_end=True,
    seed=seed,
    # metric_for_best_model=metric_name,
)

trainer = Trainer(
    model,
    args,
    train_dataset=ds_train,
    eval_dataset=ds_valid,
    tokenizer=tokenizer,
    compute_metrics=compute_metrics,
    optimizers=(AdamW(model.parameters(), lr=lr), None),  # Optimizer, Scheduler
    # compute_loss to override one must create cutsom trainer class and override the method compute_loss(self, model, inputs, return_outputs=False):
    callbacks=[EarlyStoppingCallback(early_stopping_patience=2)],
)

test_results = trainer.predict(ds_test)

test_results.metrics

calculate_metrics(test.label.values, np.argmax(test_results.predictions, axis=1))

trainer.train()

test_results = trainer.predict(ds_test)

test_results.metrics

calculate_metrics(test.label.values, np.argmax(test_results.predictions, axis=1))
