==> notebooks/81_Shap-ML.ipynb <==
import os
import numpy as np
import pandas as pd
import shap

from lightgbm import LGBMClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import (
    balanced_accuracy_score,
    make_scorer,
    f1_score,
    precision_score,
    recall_score,
)
from sklearn.model_selection import StratifiedKFold, cross_val_score
from sklearn.pipeline import Pipeline
from sklearn.svm import SVC
from xgboost import XGBClassifier
import ast

def get_x_y(df: pd.DataFrame) -> (pd.Series, pd.Series):
    """
    Get predictors and target from dataframe
    :param df: dataframe
    :return: predictors, target
    """
    x = df["text"]
    y = df["account.type"]
    y = np.where(y == "bot", 1, 0)
    return x, y

DATA_PATH = os.path.join("..", "data", "preprocessed_url_simple")
train = pd.read_csv(os.path.join(DATA_PATH, "train.csv"))
test = pd.read_csv(os.path.join(DATA_PATH, "test.csv"))

x_train, y_train = get_x_y(train)
x_test, y_test = get_x_y(test)

df = pd.read_csv("results/tfidf_optuna.csv")

MODELS_MAPPING = {
    "LGBM": LGBMClassifier,
    "XGB": RandomForestClassifier,
    "RF": LogisticRegression,
    "SVC": SVC,
    "LR": XGBClassifier,
}

KEYS_MAPPINGS = {
    "lgbm": "model__",
    "xgb": "model__",
    "rf": "model__",
    "svc": "model__",
    "tfidf": "encoder__",
}

NGRAMS_MAPPING = {"unigram": (1, 1), "digram": (1, 2), "trigram": (1, 3)}

_series = df.iloc[0, :]

_series

def load_model(
    metadata: pd.Series | dict,
    models_mapping: dict = MODELS_MAPPING,
    keys_mappings: dict = KEYS_MAPPINGS,
    ngrams_mapping: dict = NGRAMS_MAPPING,
) -> Pipeline:
    pipeline = Pipeline(
        [("encoder", TfidfVectorizer()), ("model", models_mapping[metadata["model"]]())]
    )

    params = ast.literal_eval(metadata["params"])
    params = {
        keys_mappings[k.split("_")[0]]
        + "_".join(k.split("_")[1:]): (
            v if v not in ngrams_mapping.keys() else ngrams_mapping[v]
        )
        for k, v in params.items()
    }
    pipeline.set_params(**params)
    return pipeline

pipeline = load_model(_series)

pipeline.fit(x_train, y_train)

encoder = pipeline.named_steps["encoder"]
model = pipeline.named_steps["model"]

explainer = shap.KernelExplainer(
    model.predict_proba, encoder.transform(x_train)
)  # model.predict for SVC, otherwise we can use predict_proba

shap_values = explainer.shap_values(encoder.transform(x_test[:5]))
